Refactor the Form color picker into its own component, remap logic
