+++ src/components/AccentColorPicker/AccentColorPicker.css
@@ -1,0 +1,23 @@
+.color-picker-wrapper {
+  flex-wrap: wrap;
+  justify-content: center;
+  align-items: center;
+  margin: 15px auto;
+  width: 51.5%;
+  height: fit-content;
+  background-color: #fff;
+  outline: 2px solid #000;
+}
+
+.selected-color-view {
+  width: 100%;
+  height: 40px;
+  margin: 0 auto 15px;
+  transition: background-color 0.2s ease;
+}
+
+.color-block {
+  width: 40px;
+  height: 40px;
+  cursor: pointer;
+}
