+++ src/components/Form/Form.css
@@ -93,7 +93,7 @@ textarea {
   /* overflow-y: scroll; */
 }
 
-.color-picker-wrapper {
+/* .color-picker-wrapper {
   flex-wrap: wrap;
   justify-content: center;
   align-items: center;
@@ -108,14 +108,13 @@ textarea {
   width: 100%;
   height: 40px;
   margin: 0 auto 15px;
-  /* outline: 2px solid #ff0000; */
 }
 
 .color-block {
   width: 40px;
   height: 40px;
   cursor: pointer;
-}
+} */
 
 input.submit-btn,
 .close-btn {
